fix warnings from np_schedulability_analysis, add basic analysis

+++ DAG-scheduling_Verucchi/np-schedulability-analysis/include/global/state.hpp
@@ -10,23 +10,22 @@
 #include "jobs.hpp"
 #include "cache.hpp"
 
-namespace NP {
+namespace NP
+{
 
-	namespace Global {
+	namespace Global
+	{
 
 		typedef std::size_t Job_index;
 		typedef std::vector<Job_index> Job_precedence_set;
 
-		template<class Time> class Schedule_state
+		template <class Time>
+		class Schedule_state
 		{
-			public:
-
+		public:
 			// initial state -- nothing yet has finished, nothing is running
 			Schedule_state(unsigned int num_processors)
-			: scheduled_jobs()
-			, num_jobs_scheduled(0)
-			, core_avail{num_processors, Interval<Time>(Time(0), Time(0))}
-			, lookup_key{0x9a9a9a9a9a9a9a9aUL}
+				: scheduled_jobs(), num_jobs_scheduled(0), core_avail{num_processors, Interval<Time>(Time(0), Time(0))}, lookup_key{0x9a9a9a9a9a9a9a9aUL}
 			{
 				assert(core_avail.size() > 0);
 			}
@@ -34,15 +33,13 @@ namespace NP {
 			// transition: new state by scheduling a job in an existing state,
 			//             by replacing a given running job.
 			Schedule_state(
-				const Schedule_state& from,
+				const Schedule_state &from,
 				Job_index j,
-				const Job_precedence_set& predecessors,
+				const Job_precedence_set &predecessors,
 				Interval<Time> start_times,
 				Interval<Time> finish_times,
 				hash_value_t key)
-			: num_jobs_scheduled(from.num_jobs_scheduled + 1)
-			, scheduled_jobs{from.scheduled_jobs, j}
-			, lookup_key{from.lookup_key ^ key}
+				: num_jobs_scheduled(from.num_jobs_scheduled + 1), scheduled_jobs{from.scheduled_jobs, j}, lookup_key{from.lookup_key ^ key}
 			{
 				auto est = start_times.min();
 				auto lst = start_times.max();
@@ -50,9 +47,9 @@ namespace NP {
 				auto lft = finish_times.max();
 
 				DM("est: " << est << std::endl
-				<< "lst: " << lst << std::endl
-				<< "eft: " << eft << std::endl
-				<< "lft: " << lft << std::endl);
+						   << "lst: " << lst << std::endl
+						   << "eft: " << eft << std::endl
+						   << "lft: " << lft << std::endl);
 
 				std::vector<Time> ca, pa;
 
@@ -60,7 +57,8 @@ namespace NP {
 				ca.push_back(lft);
 
 				// skip first element in from.core_avail
-				for (int i = 1; i < from.core_avail.size(); i++) {
+				for (int i = 1; i < from.core_avail.size(); i++)
+				{
 					pa.push_back(std::max(est, from.core_avail[i].min()));
 					ca.push_back(std::max(est, from.core_avail[i].max()));
 				}
@@ -68,18 +66,24 @@ namespace NP {
 				// update scheduled jobs
 				// keep it sorted to make it easier to merge
 				bool added_j = false;
-				for (const auto& rj : from.certain_jobs) {
+				for (const auto &rj : from.certain_jobs)
+				{
 					auto x = rj.first;
 					auto x_eft = rj.second.min();
 					auto x_lft = rj.second.max();
-					if (contains(predecessors, x)) {
-						if (lst < x_lft) {
+					if (contains(predecessors, x))
+					{
+						if (lst < x_lft)
+						{
 							auto pos = std::find(ca.begin(), ca.end(), x_lft);
 							if (pos != ca.end())
 								*pos = lst;
 						}
-					} else if (lst <= x_eft) {
-						if (!added_j && rj.first > j) {
+					}
+					else if (lst <= x_eft)
+					{
+						if (!added_j && rj.first > j)
+						{
 							// right place to add j
 							certain_jobs.emplace_back(j, finish_times);
 							added_j = true;
@@ -91,12 +95,12 @@ namespace NP {
 				if (!added_j)
 					certain_jobs.emplace_back(j, finish_times);
 
-
 				// sort in non-decreasing order
 				std::sort(pa.begin(), pa.end());
 				std::sort(ca.begin(), ca.end());
 
-				for (int i = 0; i < from.core_avail.size(); i++) {
+				for (size_t i = 0; i < from.core_avail.size(); i++)
+				{
 					DM(i << " -> " << pa[i] << ":" << ca[i] << std::endl);
 					core_avail.emplace_back(pa[i], ca[i]);
 				}
@@ -115,7 +119,7 @@ namespace NP {
 				return scheduled_jobs == other.scheduled_jobs;
 			}
 
-			bool can_merge_with(const Schedule_state<Time>& other) const
+			bool can_merge_with(const Schedule_state<Time> &other) const
 			{
 				assert(core_avail.size() == other.core_avail.size());
 
@@ -129,7 +133,7 @@ namespace NP {
 				return true;
 			}
 
-			bool try_to_merge(const Schedule_state<Time>& other)
+			bool try_to_merge(const Schedule_state<Time> &other)
 			{
 				if (!can_merge_with(other))
 					return false;
@@ -144,13 +148,16 @@ namespace NP {
 				auto it = certain_jobs.begin();
 				auto jt = other.certain_jobs.begin();
 				while (it != certain_jobs.end() &&
-				       jt != other.certain_jobs.end()) {
-					if (it->first == jt->first) {
+					   jt != other.certain_jobs.end())
+				{
+					if (it->first == jt->first)
+					{
 						// same job
 						new_cj.emplace_back(it->first, it->second | jt->second);
 						it++;
 						jt++;
-					} else if (it->first < jt->first)
+					}
+					else if (it->first < jt->first)
 						it++;
 					else
 						jt++;
@@ -176,9 +183,11 @@ namespace NP {
 
 			bool get_finish_times(Job_index j, Interval<Time> &ftimes) const
 			{
-				for (const auto& rj : certain_jobs) {
+				for (const auto &rj : certain_jobs)
+				{
 					// check index
-					if (j == rj.first) {
+					if (j == rj.first)
+					{
 						ftimes = rj.second;
 						return true;
 					}
@@ -197,7 +206,7 @@ namespace NP {
 				return !scheduled_jobs.contains(j);
 			}
 
-			const bool job_ready(const Job_precedence_set& predecessors) const
+			const bool job_ready(const Job_precedence_set &predecessors) const
 			{
 				for (auto j : predecessors)
 					if (!scheduled_jobs.contains(j))
@@ -205,41 +214,41 @@ namespace NP {
 				return true;
 			}
 
-			friend std::ostream& operator<< (std::ostream& stream,
-			                                 const Schedule_state<Time>& s)
+			friend std::ostream &operator<<(std::ostream &stream,
+											const Schedule_state<Time> &s)
 			{
 				stream << "Global::State(";
-				for (const auto& a : s.core_avail)
+				for (const auto &a : s.core_avail)
 					stream << "[" << a.from() << ", " << a.until() << "] ";
 				stream << "(";
-				for (const auto& rj : s.certain_jobs)
+				for (const auto &rj : s.certain_jobs)
 					stream << rj.first << "";
 				stream << ") " << s.scheduled_jobs << ")";
 				stream << " @ " << &s;
 				return stream;
 			}
 
-			void print_vertex_label(std::ostream& out,
-				const typename Job<Time>::Job_set& jobs) const
+			void print_vertex_label(std::ostream &out,
+									const typename Job<Time>::Job_set &jobs) const
 			{
-				for (const auto& a : core_avail)
+				for (const auto &a : core_avail)
 					out << "[" << a.from() << ", " << a.until() << "] ";
 				out << "\\n";
 				bool first = true;
 				out << "{";
-				for (const auto& rj : certain_jobs) {
+				for (const auto &rj : certain_jobs)
+				{
 					if (!first)
 						out << ", ";
 					out << "T" << jobs[rj.first].get_task_id()
-					    << "J" << jobs[rj.first].get_job_id() << ":"
-					    << rj.second.min() << "-" << rj.second.max();
+						<< "J" << jobs[rj.first].get_job_id() << ":"
+						<< rj.second.min() << "-" << rj.second.max();
 					first = false;
 				}
 				out << "}";
 			}
 
-			private:
-
+		private:
 			const unsigned int num_jobs_scheduled;
 
 			// set of jobs that have been dispatched (may still be running)
@@ -254,7 +263,7 @@ namespace NP {
 			const hash_value_t lookup_key;
 
 			// no accidental copies
-			Schedule_state(const Schedule_state& origin)  = delete;
+			Schedule_state(const Schedule_state &origin) = delete;
 		};
 
 	}
